Extract SCORStreamingAnalytics extension
