+++ Sources/Analytics/ComScore/SCORStreamingAnalytics.swift
@@ -1,0 +1,49 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import ComScore
+import CoreMedia
+import Player
+
+extension SCORStreamingAnalytics {
+    private static func duration(for player: Player) -> Int {
+        player.timeRange.isValid ? Int(player.timeRange.duration.seconds.toMilliseconds) : 0
+    }
+
+    private static func position(for player: Player) -> Int {
+        player.time.isValid ? Int(player.time.seconds.toMilliseconds) : 0
+    }
+
+    private static func offset(for player: Player) -> Int {
+        guard player.timeRange.isValid, player.time.isValid else { return 0 }
+        let offset = player.timeRange.end - player.time
+        return Int(offset.seconds.toMilliseconds)
+    }
+
+    func notifyEvent(for playbackState: PlaybackState, at rate: Float) {
+        switch playbackState {
+        case .playing:
+            notifyChangePlaybackRate(rate)
+            notifyPlay()
+        case .paused:
+            notifyPause()
+        case .ended:
+            notifyEnd()
+        default:
+            break
+        }
+    }
+
+    func setProperties(for player: Player, streamType: StreamType) {
+        if streamType == .dvr {
+            start(fromDvrWindowOffset: Self.offset(for: player))
+            setDVRWindowLength(Self.duration(for: player))
+        }
+        else {
+            start(fromPosition: Self.position(for: player))
+        }
+    }
+}
